skycastle-backend/requirements.txt
# conda env export --from-history
werkzeug
flask
pyopenssl
requests
oauthlib
flask-login
flask-cors
google-auth